pyproject.toml
[project]
name = "mongo-schematic"
version = "0.1.2"
description = "MongoSchematic CLI for MongoDB schema analysis, drift detection, and migrations"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Abdulazeez Abdulazeez Adeshina", email = "abdulazeez@thecodecafe.co" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

dependencies = [
  "typer[all]>=0.12.0",
  "rich>=13.7.0",
  "beanie>=1.25.0",
  "motor>=3.4.0",
  "pymongo>=4.6.1",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.1",
  "jsonschema>=4.20.0",
  "google-generativeai>=0.3.2",
  "python-dotenv>=1.0.1",
  "Faker>=24.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
mschema = "mongo_schematic.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
